Changed Div to return remainder. Moved templates for x86 code generators to separate files. More stuff for PsfPlayer.
git-svn-id: http://svn.purei.org/codegen/trunk@15 15a10f38-2c35-455d-a993-f4d4bf960095

+++ src/Jitter_CodeGen_Arm.cpp
@@ -104,7 +104,7 @@ void CCodeGen_Arm::Emit_Shift_RegRegCst(const STATEMENT& statement)
 	
 	m_assembler.Mov(g_registers[dst->m_valueLow], 
 					CArmAssembler::MakeRegisterAluOperand(g_registers[src1->m_valueLow], 
-														  CArmAssembler::MakeConstantShift(shiftType, src2->m_valueLow)));
+														  CArmAssembler::MakeConstantShift(shiftType, static_cast<uint8>(src2->m_valueLow))));
 }
 
 #define ALU_CONST_MATCHERS(ALUOP_CST, ALUOP) \
